+++ src/goveeReading.ts
@@ -1,0 +1,10 @@
+export type GoveeReading = {
+    uuid: string;
+    address: string;
+    model: string;
+    tempInC: number;
+    tempInF: number;
+    humidity: number;
+    battery: number;
+    rssi: number;
+};
